Optimize Perlin noise hot path: unchecked indexing, FMA, forced inlining
- Use unsafe get_unchecked on all permutation/gradient table lookups
  (24 bounds checks eliminated per sample, ~960 per terrain evaluation)
- Replace arithmetic with f32::mul_add (FMA) for dot products, fade
  curves, and trilinear interpolation
- Force-inline sample/sample_and_lerp/get/maintain_precission
- Work in usize throughout permutation hashing to avoid i32 casts
- Remove redundant `as f32` cast in OldBlendedNoise

Benchmark: 478ms → 415ms wall time (-13.2%), 922 → 1061 chunks/sec (+15.1%)

+++ crates/mcrs_minecraft_worldgen/src/density_function/mod.rs
@@ -2870,10 +2870,9 @@ impl DensityFunction for OldBlendedNoise {
                     xx,
                     yy,
                     zz,
-                    (self.main_smear * factor),
-                    (factored_y * factor),
-                ) as f32
-                    / factor;
+                    self.main_smear * factor,
+                    factored_y * factor,
+                ) / factor;
             }
             factor /= 2.0;
         }
